rough draft Iteration 3 working on bugs and refactoring
