+++ spec/game_spec.rb
@@ -1,0 +1,6 @@
+@player_board = Board.new
+@comp_board = Board.new
+@player_cruiser = Ship.new("Cruiser", 3)
+@player_submarine = Ship.new("Submarine", 2)
+@comp_cruiser = Ship.new("Cruiser", 3)
+@comp_submarine = Ship.new("Submarine", 2)
